refactor - comments files separated
