+++ src/components/Comment/index.js
@@ -1,1 +1,3 @@
 export { default as CommentEdit } from './CommentEdit'
+export { default as CommentsWrapper } from './Comments'
+export { default as Comment } from './Comment'
