+++ src/components/Reusable/index.js
@@ -6,5 +6,3 @@ export { default as CreatePost } from './CreatePost'
 export { default as PageHeading } from './PageHeading'
 export { default as ModalWrapper } from './Modal'
 export { default as CreateComment } from './CreateComment'
-export { default as CommentsWrapper } from './Comments'
-export { default as Comment } from './Comment'
