+++ src/pages/Post.jsx
@@ -1,10 +1,10 @@
-import { Comment, CommentsWrapper, FeedPageWrapper, ModalWrapper, Post, PostsWrapper } from "components/Reusable"
+import { FeedPageWrapper, ModalWrapper, Post } from "components/Reusable"
 import { useComments, useModal, usePosts } from "contexts"
 import { useEffect, useState } from "react"
 import { useParams } from "react-router-dom"
 import styles from 'components/Reusable/feedpage.module.css'
 import { CreateComment } from "components/Reusable"
-import { CommentEdit } from "components/Comment"
+import { CommentEdit, Comment, CommentsWrapper } from "components/Comment"
 
 const CommentModal = ModalWrapper(CommentEdit)
